:recycle: [desktop-dev/Android]  JMM 类型的统一，修复了defineJsonResponse的范型问题

+++ android/BrowserUI/src/main/java/org/dweb_browser/browserUI/download/DwebBrowserService.kt
@@ -77,7 +77,7 @@ class DwebBrowserService : Service() {
         ApiService.instance.downloadAndSave(
           path = downLoadInfo.url,
           file = File(downLoadInfo.path),
-          downLoadInfo.metaData.bundle_size.toLong(),
+          downLoadInfo.metaData.bundle_size,
           isStop = {
             when (downLoadInfo.downLoadStatus) {
               DownLoadStatus.PAUSE -> {
